% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{opt}
\alias{opt}
\title{Optimal P-value combination Test (OPT)}
\usage{
opt(p, lower.alpha = 0.01, upper.alpha = 1e+08, rep = 10000)
}
\arguments{
\item{p}{p-values to be combined}

\item{lower.alpha}{minimum alpha value considered in the test}

\item{upper.alpha}{maximum alpha value considered in the test}

\item{rep}{number of replicates to estimate the p-value using resampling}
}
\value{
the estimated test statistic, the alpha and the c values, the p-value of the test
}
\description{
Optimal P-value combination Test (OPT)
}
\examples{
p=c(0.1,0.2,0.3)
opt(p,rep=1e3)

}
