% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_test.R
\name{opt_test}
\alias{opt_test}
\title{OPT test}
\usage{
opt_test(p, l.alpha = 0.01, u.alpha = 1e+08)
}
\arguments{
\item{p}{p-values to be combined}

\item{l.alpha}{lower.alpha minimum alpha value considered in the test}

\item{u.alpha}{maximum alpha value considered in the test}
}
\value{
the estimated test statistic, the alpha and the c values
}
\description{
OPT test
}
\examples{
p=c(0.1,0.2,0.3)
opt_test(p)




}
