% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_function}
\alias{view_function}
\title{View a function's code}
\usage{
view_function(x)
}
\arguments{
\item{x}{a function}
}
\description{
A wrapper around View so that it can be used with {cli} hyperlinks
}
\keyword{internal}
